Number Wizard
Developed in a course teaching video game development in C#, this game prints messages to the console and uses binary search to narrow down on a number determined by the user at the start of the game. The range can be customized at the top of the code. If it were to be set to 1 billion, the computer can determine the correct number in about thirty turns.